requirements.txt
Django==2.1.7
djangorestframework==3.9.1
psycopg2-binary==2.7.7
pytz==2018.9
whitenoise==3.3.1
